+++ examples/rest/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.10](https://github.com/HPInc/davinci/compare/@davinci/example-rest@0.15.9...@davinci/example-rest@0.15.10) (2022-03-24)
+
+
+### Bug Fixes
+
+* use latest createRouter in examples and update doc ([4a9c56b](https://github.com/HPInc/davinci/commit/4a9c56b4e59898e1a9b5d0955e8ede7aefd4c87a))
+
+
+
+
+
+## [0.15.9](https://github.com/HPInc/davinci/compare/@davinci/example-rest@0.15.8...@davinci/example-rest@0.15.9) (2022-01-27)
+
+**Note:** Version bump only for package @davinci/example-rest
+
+
+
+
+
+## [0.15.8](https://github.com/HPInc/davinci/compare/@davinci/example-rest@0.15.7...@davinci/example-rest@0.15.8) (2022-01-21)
+
+
+### Bug Fixes
+
+* update mongoose ([e738307](https://github.com/HPInc/davinci/commit/e7383075fb9482c1bc408decab3c53682fc60342))
+* update swagger-ui-dist ([1b0fc99](https://github.com/HPInc/davinci/commit/1b0fc994fbfcd64ef3a0ffbe8c795feab48631d8))
+
+
+
+
+
 ## [0.15.7](https://github.com/HPInc/davinci/compare/@davinci/example-rest@0.15.6...@davinci/example-rest@0.15.7) (2021-11-19)
 
 **Note:** Version bump only for package @davinci/example-rest