+++ packages/mongoose/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.10](https://github.com/HPInc/davinci/compare/@davinci/mongoose@0.17.9...@davinci/mongoose@0.17.10) (2022-03-24)
+
+**Note:** Version bump only for package @davinci/mongoose
+
+
+
+
+
+## [0.17.9](https://github.com/HPInc/davinci/compare/@davinci/mongoose@0.17.8...@davinci/mongoose@0.17.9) (2022-01-27)
+
+
+### Bug Fixes
+
+* virtual as property decorator ([c53170f](https://github.com/HPInc/davinci/commit/c53170f2eb97fb1485af895b1599ee641c37cb17))
+
+
+
+
+
+## [0.17.8](https://github.com/HPInc/davinci/compare/@davinci/mongoose@0.17.7...@davinci/mongoose@0.17.8) (2022-01-21)
+
+
+### Bug Fixes
+
+* update mongoose ([e738307](https://github.com/HPInc/davinci/commit/e7383075fb9482c1bc408decab3c53682fc60342))
+
+
+
+
+
 ## [0.17.7](https://github.com/HPInc/davinci/compare/@davinci/mongoose@0.17.6...@davinci/mongoose@0.17.7) (2021-11-19)
 
 **Note:** Version bump only for package @davinci/mongoose